MB-47414 Change exit codes for command/flag parsing errors
Change the exit codes of the cbflag library in the cases of
explicit command/flag parsing errors (unknown flag, a required
flag is not specified, etc.) to a non-zero value (64).

In the case where no arguments were provided we still exit with
0 exit code and print help.

Includes unit tests for all relevant execution scenarios that
confirm that the expected exit codes are returned in all cases.

Change-Id: Ib6047daf9bba076f98be22efe51de25e126aa180
Reviewed-on: http://review.couchbase.org/c/cbflag/+/164286
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Maksimiljans Januska <maks.januska@couchbase.com>