MB-68614 Accept option value, even if it looks like a flag
Previous given an input like `--password -this-looks-like-a-flag` we
would error and say password did not have a value. This causes problems
if the value you want to supply to the option starts with a dash, e.g. a
randomly generated password.

Instead we always treat the argument after an option as the option's
value.

NOTE: This is a breaking change.

Change-Id: I28aa4410300cbc548a41225764621f5a4daa544d
Reviewed-on: https://review.couchbase.org/c/cbflag/+/234158
Tested-by: Matt Hall <matt.hall@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>